% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lat_conversion.R
\name{lat_conversion}
\alias{lat_conversion}
\title{Latitudinal coordinate conversion}
\usage{
lat_conversion(coord_lat)
}
\arguments{
\item{coord_lat}{numerical latitude to input}
}
\value{
numerical latitude in grid format
}
\description{
Converts a given latitude into the temperature grid format.
}
\examples{
lat_conversion(31.93774)

}
