% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lon_conversion.R
\name{lon_conversion}
\alias{lon_conversion}
\title{Longitudinal coordinate conversion}
\usage{
lon_conversion(coord_lon)
}
\arguments{
\item{coord_lon}{numerical longitude to input}
}
\value{
numerical longitude in grid format
}
\description{
Converts a given longitude into the temperature grid format.
}
\examples{
lon_conversion(-109.08013)

}
